悬赏0.7版本

+++ foreground/src/main/java/com/pyteam/foreground/controller/Ae07Controller.java
@@ -10,8 +10,13 @@ import org.springframework.ui.Model;
 import org.springframework.web.bind.WebDataBinder;
 import org.springframework.web.bind.annotation.*;
 
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
 import java.text.SimpleDateFormat;
 import java.util.Date;
+
+import static com.pyteam.foreground.controller.LoginController.getCookies;
+import static com.pyteam.foreground.controller.LoginController.isLogin;
 
 /**
  * @author wjm
@@ -27,41 +32,54 @@ public class Ae07Controller
     @Autowired
     private Ae02Service ae02Service;
 
+    private Boolean isLogin;
 
-    @GetMapping("/ad01/myOrder")
-    public String view(Model m)
+    @GetMapping("myOrder")
+    public String view(Model m, HttpServletRequest request, HttpServletResponse response)
     {
-        int id =1;
-        m.addAttribute("ac05",ae07Service.select(id));
-        return"/ad01/myOrder";
+        isLogin = isLogin(request, response);
+        if(isLogin)
+        {
+            m.addAttribute("isLogin",isLogin);
+            int id = Integer.parseInt(getCookies(request, "userId"));
+            m.addAttribute("ac05", ae07Service.select(id));
+            return "myOrder";
+        }
+        m.addAttribute("isLogin",isLogin);
+        return "ad01list";
     }
 
-    @PostMapping("/ad01/myOrder")
-    public String edit(@ModelAttribute("ac05") Ac05 ac05,@RequestParam("aac502") String aac502,String action)
+    @PostMapping("myOrder")
+    public String edit(@ModelAttribute("ac05") Ac05 ac05,@RequestParam("aac502") String aac502,Model m,String action,HttpServletRequest request, HttpServletResponse response)
     {
-        if(action.equals("修改订单"))
+        isLogin = isLogin(request, response);
+        if(isLogin)
         {
-            System.out.println(ac05);
-            ae07Service.update(ac05);
+            m.addAttribute("isLogin", isLogin);
 
-        }
-        else if(action.equals("取消订单"))
-        {
-            System.out.println(aac502);
-            Ac05 ac05list=new Ac05();
-            ac05list.setAac502(aac502);
-            ae07Service.delete(ac05list);
-            System.out.println("取消运行");
+            if (action.equals("修改订单"))
+            {
+                System.out.println(ac05);
+                ae07Service.update(ac05);
 
+            } else if (action.equals("取消订单"))
+            {
+                System.out.println(aac502);
+                Ac05 ac05list = new Ac05();
+                ac05list.setAac502(aac502);
+                ae07Service.delete(ac05list);
+                System.out.println("取消运行");
+
+            } else if (action.equals("完成订单"))
+            {
+                ae02Service.updatecc(aac502);
+            }
         }
-        else if(action.equals("完成订单"))
-        {
-            ae02Service.updatecc(aac502);
-        }
-        return "/ad01/myOrder";
+        m.addAttribute("isLogin",isLogin);
+        return "myOrder";
     }
 
-    @GetMapping("/ad01/delorder")
+    @GetMapping("delorder")
     public void deleteOrder(@RequestParam("aac502")String aac502)
     {
         Ac05 ac05=new Ac05();
